+++ src/modules/User/Mutation/index.ts
@@ -1,2 +1,3 @@
 export { getOrCreateUser } from "./getOrCreateUser"
+export { updateAvatarUrl } from "./updateAvatarUrl"
 export { updateUsername } from "./updateUsername"