+++ src/types.d.ts
@@ -106,16 +106,9 @@ export type User = Base & {
 export type Query = {
   __typename?: "Query"
   /** 🔒 This field requires you to be authenticated */
-  s3GetSignedPutObjectUrl: S3Payload
-  /** 🔒 This field requires you to be authenticated */
   getOrCreateUser: User
   /** 🔒 This field requires you to be authenticated */
   getUsers?: Maybe<Array<Maybe<User>>>
-}
-
-export type QueryS3GetSignedPutObjectUrlArgs = {
-  fileName: Scalars["String"]
-  fileType: Scalars["String"]
 }
 
 export type QueryGetOrCreateUserArgs = {
@@ -128,12 +121,21 @@ export type QueryGetUsersArgs = {
 
 export type Mutation = {
   __typename?: "Mutation"
+  /** 🔒 This field requires you to be authenticated */
+  s3GetSignedPutObjectUrl: S3Payload
   /** Trade a code—appended by the Cognito Hosted UI—for Cognito Tokens */
   getToken?: Maybe<AuthResponse>
   /** 🔒 This field requires you to be authenticated */
   getOrCreateUser: User
   /** 🔒 This field requires you to be authenticated */
   updateUsername: User
+  /** 🔒 This field requires you to be authenticated */
+  updateAvatarUrl: User
+}
+
+export type MutationS3GetSignedPutObjectUrlArgs = {
+  fileName: Scalars["String"]
+  fileType: Scalars["String"]
 }
 
 export type MutationGetTokenArgs = {
@@ -147,8 +149,13 @@ export type MutationGetOrCreateUserArgs = {
 }
 
 export type MutationUpdateUsernameArgs = {
-  id: Scalars["ID"]
+  id: Scalars["String"]
   username: Scalars["String"]
+}
+
+export type MutationUpdateAvatarUrlArgs = {
+  id: Scalars["String"]
+  avatarUrl: Scalars["String"]
 }
 
 /** Implemented by every 'row' in the Dynamo table */