+++ CHANGELOG.md
@@ -5,6 +5,21 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v0.6.0] - 2020-12-06
+
+### Added
+
+- updateAvatarUrl Mutation
+
+### Changed
+
+- s3GetSignedPutObjectUrl from Query to Mutation
+- Mutation.updateUsername param type
+
+### Removed
+
+- getOrCreateUser dead SK conditionExpression
+
 ## [v0.5.1] - 2020-12-01
 
 ### Fixed
@@ -200,6 +215,7 @@ add skip/take to getCommentsByUrl
 
 ### Added
 
+[v0.6.0]: https://github.com/thiskevinwang/rds-ts-node-server/compare/v0.5.1...v0.6.0
 [v0.5.1]: https://github.com/thiskevinwang/rds-ts-node-server/compare/v0.5.0...v0.5.1
 [v0.5.0]: https://github.com/thiskevinwang/rds-ts-node-server/compare/v0.4.0...v0.5.0
 [v0.4.0]: https://github.com/thiskevinwang/rds-ts-node-server/compare/v0.3.1...v0.4.0